requirements.txt
# AI ve NLP Kütüphaneleri
scikit-learn==1.0.2
nltk==3.6.7
sentence-transformers==2.2.2
faiss-cpu==1.7.2
transformers==4.16.2

# Telegram Entegrasyonu
python-telegram-bot==20.7

# Veri İşleme
pandas==1.3.5
scipy==1.7.3

# Diğer Gereksinimler
uuid==1.30
python-dateutil==2.8.2

# Geliştirme Araçları
pytest==6.2.5
black==22.1.0
mypy==0.931

# Core Dependencies
discord.py
requests
httpx
beautifulsoup4
duckduckgo-search
python-dotenv
groq
aiohttp
tabulate
termcolor
click

# Optional Dependencies
primp
colorama
